requirements.txt
numpy
pandas
streamlit
pillow